# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Booking Platformasi ucun Backend API. Randevu vaxtlari is qrafiki, nahar fasilesi ve slot addimi esasinda runtime-da hesablanir; bildirisler WebSocket ve FCM ile realtime catdirilir.",
        "title": "Booking Service API",
        "contact": {},
        "version": "2.0"
    },
    "host": "booking-service-sld9.onrender.com",
    "basePath": "/api/v1",
    "paths": {
...
        "/availability": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Isci qrafiki, nahar fasilesi, slot addimi ve movcud bronlar esasinda hesablanmis bos vaxtlar.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Availability"
                ],
                "summary": "Bos vaxtlari getir",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Isci ID",
                        "name": "staff_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Xidmet ID (mueddet bundan goturulur)",
                        "name": "service_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Baslangic tarixi YYYY-MM-DD",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Bitis tarixi YYYY-MM-DD",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_availability.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_availability.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/availability/settings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Availability"
                ],
                "summary": "Qrafik ayarlarini getir",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Isci ID (bos olsa biznes default-u)",
                        "name": "staff_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_availability.SuccessResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "slot_step_mins secim addimini teyin edir (mes. 16 deqiqe). auto_confirm true olsa bron tesdiq gozlemeden confirmed olur.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Availability"
                ],
                "summary": "Qrafik ayarlarini yenile",
                "parameters": [
                    {
                        "description": "Ayarlar",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_OrkhanNajaf1i_booking-service_internal_domain_availability.UpdateScheduleSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_availability.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/availability/time-off": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Availability"
                ],
                "summary": "Bloklanmis intervallar",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Isci ID",
                        "name": "staff_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "YYYY-MM-DD",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "YYYY-MM-DD",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_availability.SuccessResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Availability"
                ],
                "summary": "Bloklanmis interval elave et",
                "parameters": [
                    {
                        "description": "Interval",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_OrkhanNajaf1i_booking-service_internal_domain_availability.CreateTimeOffRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_availability.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/availability/time-off/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Availability"
                ],
                "summary": "Bloklanmis intervali sil",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Time off ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_availability.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/availability/working-hours": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Availability"
                ],
                "summary": "Iscinin heftelik qrafiki",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Isci ID",
                        "name": "staff_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_availability.SuccessResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Availability"
                ],
                "summary": "Heftelik qrafiki yaz",
                "parameters": [
                    {
                        "description": "Heftelik qrafik",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_OrkhanNajaf1i_booking-service_internal_domain_availability.BulkWorkingHoursRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_availability.SuccessResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Baslangic/bitis saati ve (istese) nahar fasilesi. break_enabled=false olsa fasile tetbiq edilmir.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Availability"
                ],
                "summary": "Bir gunun is saatlarini teyin et",
                "parameters": [
                    {
                        "description": "Is saatlari",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_OrkhanNajaf1i_booking-service_internal_domain_availability.SetWorkingHoursRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_availability.SuccessResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Availability"
                ],
                "summary": "Bir gunun qrafikini sil",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Isci ID",
                        "name": "staff_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "0=Bazar ... 6=Senbe",
                        "name": "day_of_week",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_availability.SuccessResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_OrkhanNajaf1i_booking-service_internal_domain_availability.BulkWorkingHoursRequest": {
            "type": "object",
            "properties": {
                "days": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_OrkhanNajaf1i_booking-service_internal_domain_availability.SetWorkingHoursRequest"
                    }
                },
                "staff_id": {
                    "type": "string"
                }
            }
        },
        "github_com_OrkhanNajaf1i_booking-service_internal_domain_availability.CreateTimeOffRequest": {
            "type": "object",
            "properties": {
                "end_at": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "staff_id": {
                    "type": "string"
                },
                "start_at": {
                    "type": "string"
                }
            }
        },
        "github_com_OrkhanNajaf1i_booking-service_internal_domain_availability.SetWorkingHoursRequest": {
            "type": "object",
            "properties": {
                "break_enabled": {
                    "type": "boolean"
                },
                "break_end": {
                    "type": "string"
                },
                "break_start": {
                    "type": "string"
                },
                "day_of_week": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "staff_id": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "github_com_OrkhanNajaf1i_booking-service_internal_domain_availability.UpdateScheduleSettingsRequest": {
            "type": "object",
            "properties": {
                "allow_reschedule_proposal": {
                    "type": "boolean"
                },
                "auto_confirm": {
                    "type": "boolean"
                },
                "buffer_after_mins": {
                    "type": "integer"
                },
                "buffer_before_mins": {
                    "type": "integer"
                },
                "default_duration_mins": {
                    "type": "integer"
                },
                "max_advance_days": {
                    "type": "integer"
                },
                "min_notice_mins": {
                    "type": "integer"
                },
                "slot_step_mins": {
                    "type": "integer"
                },
                "staff_id": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
...
        "internal_http_handlers_availability.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "internal_http_handlers_availability.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}